refactor: free post directory structor
