+++ src/apis/free-posts/constants/free-post-comment.constant.ts
@@ -1,0 +1,4 @@
+export const FREE_POST_COMMENT_DESCRIPTION_LENGTH = {
+  MIN: 1,
+  MAX: 255,
+} as const;
