+++ src/apis/free-posts/constants/free-post-comment.enum.ts
@@ -1,0 +1,4 @@
+export enum FreePostCommentStatus {
+  Posting = 'posting',
+  Remove = 'remove',
+}
