+++ src/apis/free-posts/constants/free-post-reply-comment.constant.ts
@@ -1,8 +1,3 @@
-export const FREE_POST_COMMENT_DESCRIPTION_LENGTH = {
-  MIN: 1,
-  MAX: 255,
-} as const;
-
 export const FREE_POST_REPLY_COMMENT_DESCRIPTION_LENGTH = {
   MIN: 1,
   MAX: 255,
