+++ src/apis/free-posts/constants/free-post-reply-comment.enum.ts
@@ -1,8 +1,3 @@
-export enum FreePostCommentStatus {
-  Posting = 'posting',
-  Remove = 'remove',
-}
-
 export enum FreePostReplyCommentStatus {
   Posting = 'posting',
   Remove = 'remove',
