+++ src/apis/free-posts/dto/find-free-post-comment-list-query.dto.ts
@@ -1,15 +1,15 @@
+import { FreePostCommentStatus } from '@src/apis/free-posts/constants/free-post-comment.enum';
 import { FREE_POST_ORDER_FIELD } from '@src/apis/free-posts/constants/free-post.constant';
-import { FreePostReplyCommentStatus } from '@src/apis/free-posts/free-post-comments/constants/free-post-comment.enum';
-import { FreePostReplyCommentDto } from '@src/apis/free-posts/free-post-comments/dto/free-post-reply-comment.dto';
+import { FreePostCommentDto } from '@src/apis/free-posts/dto/free-post-comment.dto';
 import { SortOrder } from '@src/constants/enum';
 import { PageDto } from '@src/dto/page.dto';
 import { ApiPropertyOrder } from '@src/dto/swagger/api-property-order.decorator';
 import { CsvToOrder, Order } from '@src/dto/transformer/csv-to-order.decorator';
 import { IsDefined, IsOptional } from 'class-validator';
 
-export class FindFreePostReplyCommentListQueryDto
+export class FindFreePostCommentListQueryDto
   extends PageDto
-  implements Partial<FreePostReplyCommentDto>
+  implements Partial<FreePostCommentDto>
 {
   @ApiPropertyOrder(FREE_POST_ORDER_FIELD)
   @CsvToOrder<typeof FREE_POST_ORDER_FIELD>([...FREE_POST_ORDER_FIELD])
@@ -17,5 +17,5 @@ export class FindFreePostReplyCommentListQueryDto
   order: Order<typeof FREE_POST_ORDER_FIELD> = { id: SortOrder.Desc };
 
   @IsDefined()
-  status: FreePostReplyCommentStatus = FreePostReplyCommentStatus.Posting;
+  status: FreePostCommentStatus = FreePostCommentStatus.Posting;
 }
