+++ src/apis/free-posts/dto/find-free-post-reply-comment-list-query.dto.ts
@@ -1,15 +1,15 @@
 import { FREE_POST_ORDER_FIELD } from '@src/apis/free-posts/constants/free-post.constant';
-import { FreePostCommentStatus } from '@src/apis/free-posts/free-post-comments/constants/free-post-comment.enum';
-import { FreePostCommentDto } from '@src/apis/free-posts/free-post-comments/dto/free-post-comment.dto';
+import { FreePostReplyCommentDto } from '@src/apis/free-posts/dto/free-post-reply-comment.dto';
 import { SortOrder } from '@src/constants/enum';
 import { PageDto } from '@src/dto/page.dto';
 import { ApiPropertyOrder } from '@src/dto/swagger/api-property-order.decorator';
 import { CsvToOrder, Order } from '@src/dto/transformer/csv-to-order.decorator';
 import { IsDefined, IsOptional } from 'class-validator';
+import { FreePostReplyCommentStatus } from '../constants/free-post-reply-comment.enum';
 
-export class FindFreePostCommentListQueryDto
+export class FindFreePostReplyCommentListQueryDto
   extends PageDto
-  implements Partial<FreePostCommentDto>
+  implements Partial<FreePostReplyCommentDto>
 {
   @ApiPropertyOrder(FREE_POST_ORDER_FIELD)
   @CsvToOrder<typeof FREE_POST_ORDER_FIELD>([...FREE_POST_ORDER_FIELD])
@@ -17,5 +17,5 @@ export class FindFreePostCommentListQueryDto
   order: Order<typeof FREE_POST_ORDER_FIELD> = { id: SortOrder.Desc };
 
   @IsDefined()
-  status: FreePostCommentStatus = FreePostCommentStatus.Posting;
+  status: FreePostReplyCommentStatus = FreePostReplyCommentStatus.Posting;
 }
