+++ src/apis/free-posts/dto/free-post-comments-item.dto.ts
@@ -1,0 +1,3 @@
+import { FreePostCommentDto } from '@src/apis/free-posts/dto/free-post-comment.dto';
+
+export class FreePostCommentsItemDto extends FreePostCommentDto {}
