+++ src/apis/free-posts/free-post-history/repositories/free-post-comment-history.repository.ts
