+++ src/apis/free-posts/free-post-history/services/free-post-history.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
-import { FreePostCommentHistoryRepository } from '@src/apis/free-posts/free-post-comments/repositories/free-post-comment-history.repository';
-import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-posts/free-post-comments/repositories/free-post-reply-comment-history.repository';
 import { CreateFreePostCommentHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-comment-history.dto';
 import { CreateFreePostHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-history.dto';
 import { CreateFreePostReplyCommentHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-reply-comment-history.dto';
+import { FreePostCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-comment-history.repository';
 import { FreePostHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-history.repository';
+import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-reply-comment-history.repository';
 import { HistoryAction, SortOrder } from '@src/constants/enum';
 import { EntityManager } from 'typeorm';
